bug fix: course not being deleted from database fail-safe
Sometimes there is a glitch where a user will delete a review - and it's the only review for that course, but the course won't be deleted from the database even after the review is deleted. Added a fail-safe where a user will be able to post a review to these courses, even though the "beingUpdated" value is set to another userid.